+++ plone/app/tiles/drafting.py
@@ -1,9 +1,9 @@
 # -*- coding: utf-8 -*-
 from plone.app.drafts.interfaces import ICurrentDraftManagement
-from plone.app.drafts.interfaces import USERID_KEY
 from plone.app.drafts.interfaces import IDraft
-from plone.app.drafts.interfaces import IDraftSyncer
 from plone.app.drafts.interfaces import IDrafting
+from plone.app.drafts.interfaces import IDraftSyncer
+from plone.app.drafts.interfaces import USERID_KEY
 from plone.app.drafts.proxy import DraftProxy
 from plone.app.drafts.utils import getCurrentDraft
 from plone.app.tiles.interfaces import ITilesFormLayer
@@ -14,9 +14,9 @@ from urlparse import urlparse
 from zope.annotation.interfaces import IAnnotations
 from zope.component import adapter
 from zope.component import adapts
-from zope.interface import Interface
 from zope.interface import implementer
 from zope.interface import implements
+from zope.interface import Interface
 
 try:
     from plone.app.drafts.dexterity import IDisplayFormDrafting